+++ forge.config.ts
@@ -148,7 +148,6 @@ const config: ForgeConfig = {
           config: 'vite.renderer.config.ts',
         },
       ],
-      concurrent: false,
     }),
     // Fuses are used to enable/disable various Electron functionality
     // at package time, before code signing the application