+++ src/codegen/browser/selectors.ts
@@ -1,60 +1,19 @@
-import { AriaRole } from 'react'
-
 import { ElementSelector } from '@/schemas/recording'
+import {
+  GetByAltTextNodeSelector,
+  GetByLabelNodeSelector,
+  GetByRoleNodeSelector,
+  GetByPlaceholderNodeSelector,
+  GetByTitleNodeSelector,
+  GetByTestIdNodeSelector,
+  CssNodeSelector,
+  NodeSelector,
+} from '@/schemas/selectors'
 import { exhaustive } from '@/utils/typescript'
 
-export interface CssSelector {
-  type: 'css'
-  selector: string
-}
-
-export interface GetByRoleSelector {
-  type: 'role'
-  role: AriaRole
-  name: string
-}
-
-export interface GetByAltTextSelector {
-  type: 'alt'
-  text: string
-}
-
-export interface GetByLabelSelector {
-  type: 'label'
-  text: string
-}
-
-export interface GetByPlaceholderSelector {
-  type: 'placeholder'
-  text: string
-}
-
-export interface GetByTextSelector {
-  type: 'text'
-  text: string
-}
-
-export interface GetByTitleSelector {
-  type: 'title'
-  text: string
-}
-
-export interface GetByTestIdSelector {
-  type: 'test-id'
-  testId: string
-}
-
-export type NodeSelector =
-  | CssSelector
-  | GetByRoleSelector
-  | GetByAltTextSelector
-  | GetByLabelSelector
-  | GetByPlaceholderSelector
-  | GetByTextSelector
-  | GetByTitleSelector
-  | GetByTestIdSelector
-
-function getRoleSelector(selectors: ElementSelector): GetByRoleSelector | null {
+function getRoleSelector(
+  selectors: ElementSelector
+): GetByRoleNodeSelector | null {
   if (selectors.role === undefined) {
     return null
   }
@@ -68,7 +27,7 @@ function getRoleSelector(selectors: ElementSelector): GetByRoleSelector | null {
 
 function getAltTextSelector(
   selectors: ElementSelector
-): GetByAltTextSelector | null {
+): GetByAltTextNodeSelector | null {
   if (selectors.alt === undefined) {
     return null
   }
@@ -81,7 +40,7 @@ function getAltTextSelector(
 
 function getLabelSelector(
   selectors: ElementSelector
-): GetByLabelSelector | null {
+): GetByLabelNodeSelector | null {
   if (selectors.label === undefined) {
     return null
   }
@@ -94,7 +53,7 @@ function getLabelSelector(
 
 function getPlaceholderSelector(
   selectors: ElementSelector
-): GetByPlaceholderSelector | null {
+): GetByPlaceholderNodeSelector | null {
   if (selectors.placeholder === undefined) {
     return null
   }
@@ -107,7 +66,7 @@ function getPlaceholderSelector(
 
 function getTitleSelector(
   selectors: ElementSelector
-): GetByTitleSelector | null {
+): GetByTitleNodeSelector | null {
   if (selectors.title === undefined) {
     return null
   }
@@ -120,7 +79,7 @@ function getTitleSelector(
 
 function getTestIdSelector(
   selectors: ElementSelector
-): GetByTestIdSelector | null {
+): GetByTestIdNodeSelector | null {
   if (selectors.testId === undefined || selectors.testId.trim() === '') {
     return null
   }
@@ -131,7 +90,7 @@ function getTestIdSelector(
   }
 }
 
-function getCssSelector(selectors: ElementSelector): CssSelector {
+function getCssSelector(selectors: ElementSelector): CssNodeSelector {
   return {
     type: 'css',
     selector: selectors.css,