+++ src/rules/utils.ts
@@ -64,9 +64,7 @@ export function urlToQueryParams(url: string): Query[] {
   }
 }
 
-// @ts-expect-error we have commonjs set as module option
 if (import.meta.vitest) {
-  // @ts-expect-error we have commonjs set as module option
   const { it, expect } = import.meta.vitest
 
   const generateResponse = (