+++ src/views/Validator/Browser/BrowserDebugger.tsx
@@ -1,5 +1,6 @@
 import { css } from '@emotion/react'
 import { Flex, Tabs } from '@radix-ui/themes'
+import { useState } from 'react'
 
 import { LogsSection } from '@/components/Validator/LogsSection'
 import {
@@ -9,6 +10,7 @@ import {
   useDefaultLayout,
   usePanelCallbackRef,
 } from '@/components/primitives/ResizablePanel'
+import { NodeSelector } from '@/schemas/selectors'
 
 import { DebugSession } from '../types'
 
@@ -27,6 +29,9 @@ export function BrowserDebugger({
   session,
   onDebugScript,
 }: BrowserDebuggerProps) {
+  const [highlightedSelector, setHighlightedSelector] =
+    useState<NodeSelector | null>(null)
+
   const [drawer, setDrawer] = usePanelCallbackRef()
 
   const drawerLayout = useDefaultLayout({
@@ -70,13 +75,18 @@ export function BrowserDebugger({
               `}
             >
               <Panel id="main" minSize={400}>
-                <BrowserOverviewPanel script={script} session={session} />
+                <BrowserOverviewPanel
+                  script={script}
+                  session={session}
+                  highlightedSelector={highlightedSelector}
+                />
               </Panel>
               <Separator />
               <Panel id="actions" minSize={400}>
                 <BrowserActionsPanel
                   session={session}
                   onDebugScript={onDebugScript}
+                  onHighlight={setHighlightedSelector}
                 />
               </Panel>
             </Group>