+++ src/views/Validator/Browser/SessionPlayer/SessionPlayer.hooks.ts
@@ -266,5 +266,6 @@ export function usePlayer({ streaming, mount, events }: UsePlayerOptions) {
     play,
     pause,
     seek,
+    player,
   }
 }